requirements.txt
Flask
gunicorn
python-dotenv
flask-sqlalchemy
flask-migrate
